requirements.txt
streamlit 
pandasai
python-dotenv